backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
psutil
requests
PyJWT
glob2
python-multipart
bcrypt>=4.0.1
passlib>=1.7.4
slowapi